correção na mensagem 'carregando'

+++ frontend/react+crud+api/src/pages/Home.jsx
@@ -13,14 +13,27 @@ import ClientList from "../components/ClientList";
 function Home() {
   const [clients, setClients] = useState([]);
   const [editingClient, setEditingClient] = useState(null);
+  //  estado de loading começando como 'true'
+  const [loading, setLoading] = useState(true);
 
   useEffect(() => {
     loadClients();
   }, []);
 
   async function loadClients() {
-    const data = await getClients();
-    setClients(data);
+    try {
+      setLoading(true);
+
+      // Força o React a respirar por 100ms antes de travar a thread com o fetch
+      // await new Promise(resolve => setTimeout(resolve, 100));
+
+      const data = await getClients();
+      setClients(data);
+    } catch (error) {
+      console.error("Erro ao carregar clientes:", error);
+    } finally {
+      setLoading(false);
+    }
   }
 
   async function handleSaveClient(client) {
@@ -30,13 +43,14 @@ function Home() {
     } else {
       await createClient(client);
     }
-
     loadClients();
   }
 
   async function handleDeleteClient(id) {
-    await deleteClient(id);
-    loadClients();
+    if (window.confirm("Deseja realmente excluir este cliente?")) {
+      await deleteClient(id);
+      loadClients();
+    }
   }
 
   function handleEditClient(client) {
@@ -52,11 +66,20 @@ function Home() {
         editingClient={editingClient}
       />
 
-      <ClientList
-        clients={clients}
-        onDeleteClient={handleDeleteClient}
-        onEditClient={handleEditClient}
-      />
+      {/* 2. Verificação: Se loading for true, mostra a mensagem. Se não, mostra a lista */}
+      {loading ? (
+        <div className="loading-container">
+          <div className="spinner"></div>
+          <p>Carregando dados...</p>
+          <small>O servidor gratuito do Render pode demorar até 1 min para acordar.</small>
+        </div>
+      ) : (
+        <ClientList
+          clients={clients}
+          onDeleteClient={handleDeleteClient}
+          onEditClient={handleEditClient}
+        />
+      )}
     </>
   );
 }
